requirements.txt
Flask==1.1.2
Flask_Bootstrap==3.3.7.1
Flask_CKEditor==0.4.6
Flask_Gravatar==0.5.0
Flask_Login==0.6.2
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.4.3
Flask_WTF==1.0.1
python-dotenv==0.21.0
requests==2.25.1
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==3.0.1
asposestorage==1.0.2
pkg==0.2
controller==0.1.0
